requirements.txt
hydra-core==1.3.2
einops==0.7.0
pypinyin==0.51.0
